requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
decorator==4.4.2
deepgram-sdk==2.11.0
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
frozenlist==1.4.0
google-api-core==2.12.0
google-auth==2.23.2
google-cloud-speech==2.21.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
inflection==0.5.1
kombu==5.3.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.24.4
packaging==23.2
Pillow==10.0.1
proglog==0.1.10
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
websockets==11.0.3
yarl==1.9.2
